// spotify-project/src/app/models/TopTracks.ts
// Generated by https://quicktype.io

export class TopTracksPagingObject {
    href: string;
    items: TopTracks[];
    limit: number;
    next: string;
    offset: number;
    previous: string;
    total: number;
}

export class TopTracks {
    items:    Item[];
    total:    number;
    limit:    number;
    offset:   number;
    previous: null;
    href:     string;
    next:     string;
}

export class Item {
    album:             Album;
    artists:           Artist[];
    available_markets: string[];
    disc_number:       number;
    duration_ms:       number;
    explicit:          boolean;
    external_ids:      ExternalIDS;
    external_urls:     ExternalUrls;
    href:              string;
    id:                string;
    is_local:          boolean;
    name:              string;
    popularity:        number;
    preview_url:       string;
    track_number:      number;
    type:              ItemType;
    uri:               string;
}

export class Album {
    album_type:             AlbumType;
    artists:                Artist[];
    available_markets:      string[];
    external_urls:          ExternalUrls;
    href:                   string;
    id:                     string;
    images:                 Image[];
    name:                   string;
    release_date:           string;
    release_date_precision: ReleaseDatePrecision;
    total_tracks:           number;
    type:                   AlbumTypeEnum;
    uri:                    string;
}

export enum AlbumType {
    Album = "ALBUM",
    Single = "SINGLE",
}

export class Artist {
    external_urls: ExternalUrls;
    href:          string;
    id:            string;
    name:          string;
    type:          ArtistType;
    uri:           string;
}

export class ExternalUrls {
    spotify: string;
}

export enum ArtistType {
    Artist = "artist",
}

export class Image {
    height: number;
    url:    string;
    width:  number;
}

export enum ReleaseDatePrecision {
    Day = "day",
    Year = "year",
}

export enum AlbumTypeEnum {
    Album = "album",
}

export class ExternalIDS {
    isrc: string;
}

export enum ItemType {
    Track = "track",
}