// spotify-project/src/app/models/AudioFeatures.ts
// Generated by https://quicktype.io
export class AudioFeaturesWrapper {
    audioFeatures: AudioFeatures[];
}

export class AudioFeatures {
    danceability:     number;
    energy:           number;
    key:              number;
    loudness:         number;
    mode:             number;
    speechiness:      number;
    acousticness:     number;
    instrumentalness: number;
    liveness:         number;
    valence:          number;
    tempo:            number;
    type:             string;
    id:               string;
    uri:              string;
    track_href:       string;
    analysis_url:     string;
    duration_ms:      number;
    time_signature:   number;
}