// spotify-project/src/app/models/TopPlaylists.ts
// Generated by https://quicktype.io

export class PlaylistPagingObject {
    href:     string;
    items:    Playlist[];
    limit:    number;
    next:     string;
    offset:   number;
    previous: null;
    total:    number;
}

export class Playlist {
    collaborative: boolean;
    external_urls: ExternalUrls;
    href:          string;
    id:            string;
    images:        Image[];
    name:          string;
    owner:         Owner;
    primary_color: null;
    public:        boolean;
    snapshot_id:   string;
    tracks:        Tracks;
    type:          ItemType;
    uri:           string;
}

export class ExternalUrls {
    spotify: string;
}

export class Image {
    height: number | null;
    url:    string;
    width:  number | null;
}

export class Owner {
    display_name:  string;
    external_urls: ExternalUrls;
    href:          string;
    id:            string;
    type:          OwnerType;
    uri:           string;
}

export enum OwnerType {
    User = "user",
}

export class Tracks {
    href:  string;
    total: number;
}

export enum ItemType {
    Playlist = "playlist",
}